aider: Update `update_release_draft` job in `release-drafter.yaml` to depend on `pre-pr-merge` job completion using `workflow_run` event.
# Aider chat conversation:

USER: `update_release_draft` appears to have a problem. It is dependent on a job from another workflow file. How do we make it depend on the pre-pr-merge job

ASSISTANT: To make the `update_release_draft` job in the `release-drafter.yaml` workflow file depend on the completion of the `pre-pr-merge` job from the `pre-pr-merge.yaml` workflow file, you would need to use the `workflow_run` event. This allows one workflow to be triggered after another workflow has completed.

Here is the diff to update the `release-drafter.yaml` workflow file to trigger after the `pre-pr-merge` workflow completes:

```diff
--- .github/workflows/release-drafter.yaml
+++ .github/workflows/release-drafter.yaml
@@ ... @@
 on:
   push:
     branches:
       - master
+  workflow_run:
+    workflows: ["Pre-PR Merge CI"]
+    types:
+      - completed
   pull_request:
     types: [opened, reopened, synchronize]

 jobs:
   update_release_draft:
-    needs: pre-pr-merge
     permissions:
       contents: write
       pull-requests: write
     runs-on: ubuntu-latest
     steps:
       - uses: release-drafter/release-drafter@v5
         with:
           config-name: release-drafter-config.yaml
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
```

This change will trigger the `update_release_draft` job when the `Pre-PR Merge CI` workflow has completed. Note that the `needs` key has been removed because it is not applicable across separate workflow files. Instead, the `workflow_run` event is used to create a dependency between workflows.